ui : 리스트 추가 팝업 디자인 수정
