+++ src/App.jsx
@@ -10,7 +10,7 @@ import TodoListItemPopup from "./components/TodoListDeletePopup";
 import TodoListSearchPopup from "./components/TodoListSearchPopup";
 import TodoDetailPopup from "./components/TodoDetailPopup";
 import { useDispatch, useSelector } from "react-redux";
-import AddTodoList from "./components/AddTodoList";
+import TodoListAddPopup from "./components/AddTodoList";
 import TodoListDeletePopup from "./components/TodoListDeletePopup";
 import { useState } from "react";
 import TodoListEditTitlePopup from "./components/TodoListEditTitlePopup";
@@ -35,7 +35,7 @@ function App() {
         </MobileFeatureLayout>
         <TodoList setSelectedId={setSelectedId} />
       </AppLayout>
-      {mode.mode === "add" ? <AddTodoList /> : null}
+      {mode.mode === "add" ? <TodoListAddPopup /> : null}
       {mode.mode === "delete" ? (
         <TodoListDeletePopup selectedId={selectedId} />
       ) : null}
